woo tests pass

+++ flocker/acceptance/testtools.py
@@ -338,25 +338,25 @@ def get_clean_nodes(test_case, num_nodes):
                 d = cluster.remove_container(container[u"name"])
                 results_list.append(d)
             deleting = gather_deferreds(results_list)
+
+            def no_containers(cluster):
+                d = cluster.current_containers()
+                d.addCallback(lambda result: len(result[1]) == 0)
+                return d
+
+            def check_containers(cluster):
+                d = loop_until(lambda: no_containers(cluster))
+                d.addCallback(lambda _: cluster)
+                return d
+
+            deleting.addCallback(lambda _: cluster)
+            deleting.addCallback(check_containers)
             deleting.addCallback(lambda _: cluster)
             return deleting
 
         get_containers.addCallback(delete_containers)
 
-        def no_containers(cluster):
-            d = cluster.current_containers()
-            d.addCallback(lambda result: len(result[1]) == 0)
-            return d
-
-        def check_clean(cluster):
-            d = loop_until(lambda: no_containers(cluster))
-            d.addCallback(lambda _: cluster)
-            return d
-
-        get_containers.addCallback(check_clean)
         return get_containers
-
-    getting.addCallback(clean_containers)
 
     def clean_datasets(cluster):
         get_datasets = cluster.configured_datasets()
@@ -364,27 +364,31 @@ def get_clean_nodes(test_case, num_nodes):
         def delete_datasets(datasets):
             results_list = []
             for dataset in datasets:
-                d = cluster.delete_dataset(dataset[u"dataset_id"])
-                results_list.append(d)
+                if not dataset[u"deleted"]:
+                    d = cluster.delete_dataset(dataset[u"dataset_id"])
+                    results_list.append(d)
             deleting = gather_deferreds(results_list)
+
+            def no_datasets(cluster):
+                d = cluster.datasets_state()
+                d.addCallback(lambda datasets: len(datasets) == 0)
+                return d
+
+            def check_datasets(cluster):
+                d = loop_until(lambda: no_datasets(cluster))
+                d.addCallback(lambda _: cluster)
+                return d
+
+            deleting.addCallback(lambda _: cluster)
+            deleting.addCallback(check_datasets)
             deleting.addCallback(lambda _: cluster)
             return deleting
 
         get_datasets.addCallback(delete_datasets)
 
-        def no_datasets(cluster):
-            d = cluster.datasets_state()
-            d.addCallback(lambda datasets: len(datasets) == 0)
-            return d
-
-        def check_datasets(cluster):
-            d = loop_until(lambda: no_datasets(cluster))
-            d.addCallback(lambda _: cluster)
-            return d
-
-        get_datasets.addCallback(check_datasets)
         return get_datasets
 
+    getting.addCallback(clean_containers)
     getting.addCallback(clean_datasets)
     getting.addCallback(lambda _: reachable_nodes)
     return getting
